correct nms

+++ src/param.py
@@ -87,7 +87,7 @@ def parse_args():
     # Training configuration
     parser.add_argument("--multiGPU", action='store_const', default=False, const=True)
     parser.add_argument("--numWorkers", dest='num_workers', default=0)
-    parser.add_argument("--nms", type=int, default=None)
+    parser.add_argument("--nms", type=float, default=None)
 
     # Parse the arguments.
     args = parser.parse_args()
